Make CartDNS use ByteCart API
Entries are now automatically created and removed using the 3rd line of
station signs as hostname. Updaters can record names automatically.

Fix a bug in DNS list preventing to run it if some permissions are not
set.

Refactor the whole code to allow calls from CommandExecutor or from
Listener.

Allow more special chars in hostnames and prevent injection of 'union'
in SQL statements.